requirements.txt
crewai
streamlit
requests
python-dotenv
litellm==1.60.2